Update script.js
Appended a fix for the sleep() function on notifications. Will work now, without errors. Also updated the sleep function altering the opacity every 333ms, reducing it by 25% each time so the notification slowly fades out until the display is set to none. Also had to re-establish the opacity back to 1.0 when the notification is called again, as if the notification is called twice without it the opacity will still be 25%.